requirements.txt
# Core
pytz==2021.1
Django==3.1.7
django-configurations==2.2
gunicorn==20.1.0
newrelic==6.2.0.156

# For the persistence stores
psycopg2-binary==2.8.6
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.1.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.12.4
Markdown==3.3.4
django-filter==2.4.0
django-extensions==3.1.3
django-cors-headers==3.7.0

# Celery==5.1.0
# django-redis==4.10.0

# Developer Tools
ipdb==0.13.7
ipython==7.22.0
mkdocs==1.1.2
flake8==3.9.0
paramiko==2.7.2
numpy==1.20.3
celery==5.1.1
redis==3.5.3

# Testing
mock==4.0.3
factory-boy==3.2.0
django-nose==1.4.7
nose-progressive==1.5.2
coverage==5.5

# Static and Media Storage
django-storages==1.11.1
boto3==1.17.43

